file search: include workspace folder in filter
It is now possible to filter file by the workspace folder in the "Go to File..." popup.

This is particularly useful when your workspace contains multiple projects each with the same file.

A common example is the `README.md` file. Say you want to find a `README.md` file from a particular folder, say `my-folder`. Before this change you'd have press the `Down` button a few times before you could get to the file.

With this change you'd instead search for `mfREADME.md`. The desired readme file should now appear closer to the top of the file search popup.